refactor: Timer hooks 디렉토리로 분리 #261
